Sorting, data structures
Signed-off-by: davidjacobo <jguillen@cimat.mx>

This can be seen as a min-max problem, lets assume we need our table
to have P legs of length equals di, thus we have to remove all the legs
of longer length than that AND we can keep an ammount P-1 of smaller legs.

There are a couple of questions here, which legs should I decide to keep?
How do I choose the length di?

You should keep the longer legs, (thus minimizing the removal of the rest)
and as you don't have idea in advance (and the range is not that bad) you
can give the chance of every di to be selected.