DP, Segment trees
Signed-off-by: davidjacobo <jguillen@cimat.mx>

First of all we need lazy propagation to mark all the updates to do
into a segment tree. After that, we know that we want to remove only
negative reviews, for that we can use dynamic programming to pick a
subset of reviews which maximize the negative reviews to remove.