Disjoint sets, binary search, dp
Signed-off-by: davidjacobo <jguillen@cimat.mx>

The basic idea is to start adding 1 element at a time to the original
array, you have to start from smaller to bigger (a min priority
queue does this in the implementation).

After that, disjoint set tells us how many consecutive cells are
already in place at each of the sides for a new cell, assume p and
q are the consecutive cells at the left and right respectively
for a cell a_i, then the ammount of arrays wich a_i is the max
elemnt is (p+1)*(q+1).

After knowing how many times each cell is the max, only sort and
perform an integral sum. Binary search over this array is piece of
cake as we have to solve queries <, >, or == to a given query K.