Binary search, implementation
Signed-off-by: davidjacobo <jguillen@cimat.mx>

As you probably noticed, a new item will stack over any item which is
strickly less than itself, from this subset, it will choose the smaller
element. This set of numbers, will always be non-decreasing, even with
the changes.

Puting dummys at the start and end allow us to evade corner cases like
the new element being inserted at the begin or end in the vector.
As per each element we have to do a binary search to find the correct
index to place it, a time complexity of ~O(N log N) suffices.