DP, bitmask
Signed-off-by: davidjacobo <jguillen@cimat.mx>

Once you realize this is a DP problem you need to specify what is a
state and how to transition thru them, a natural idea would be
to use each of the possible 2^N subsets and a reference to the last
used dish.

In memory that would look like ~O(2^N * N) when N <= 18 this is an easy
decision which needs no prunning. Calculate all the states from the very
start and you are done.