Simulation, Strings
Signed-off-by: David Jacobo <jguillen@cimat.mx>

The only detail you need to notice is that simulating the full process
can take way too long, on the other hand, is not necesary. This is
because contiguous segments of 2 or more of the same character will
never be flipped, they will spread to adjacent cells iff they are not
part of a similar group (wether those are 0's or 1's).

As every cell is entering into the structure at most once, this algorithm
is ~O(N)