DP, Graph theory
First pre-process the graph, it is a bipartite graph with (possibly)
multiple components, so detect those components (elements on the left,
right and the total cost of inverting it!)

After that you can threat the problem as a Knapsack problem, either
inverting or not each component (it is worth inverting iff there is a
gain in the number of the elements)

Cool thing about this? only 200 components at most, and the budget is
10^4; the table used to represent these states is then 2*10^6, which
also corresponds to the time complexity.