Dfs
The key to tackle this problem is to notice that the empty spaces are
part of the same component, and they must remain that way. So pick any
of the empty spaces and launch dfs with a "time to live" variable which
is equals to the amount of spaces-k. when it turns < 0 you should stop
traversing the spaces, the ones that we didnt traverse are the ones we
need to build a wall upon them (X)