Greedy
The key here is to take the costs (either vertical or horizontal, it
doesn't matter which you take first) from bigger to smaller, and keep
track of how many horizontal and vertical segments you have. O(N log N)
as we rely on a priority queue to keep track of the biggest element, any
sort will do.