Greedy
As the input is already sorted you only have to match certain elements
in the array, if the array was twice the length of the amount of boxes
you will need to match (0, n-1) , (1, n-2) and so on, but as that is not
always true you could potentially leave some items in a single box each,
which ones? the largest!