Graph theory
We will need BFS to compute distances in a tree, this will help us to
compute the diameter of a tree; the diameter of a tree can be found by
running BFS once on a tree, finding the furthest vertex, then launch BFS
again starting from that vertex.

Do this for all the possible subtrees, those can be found by removing an
edge at the time (V-1 edges).

O(V^2) can be obtained since per each partition we run the linear
algorithm BFS. We don't actually need DFS, but I assumed it would be
cool to put useless code :)