Graph theory
Keep track of the different components (nationalities) with a
disjoint-set data structure. After that each component can be related
with all the other components, so you could think something like
n*(n-1)/2 to build all the possible pairs. Summing the size of all the
components can be done in a single sweep, and after that, another single
sweep to build the total count.