Implementation
Signed-off-by: David Jacobo <jguillen@cimat.mx>

As it is ensured that n-1 numbers are just a 1 and 0 or more '0's
just identify the number which is not beautiful and count the zeros on
the guys which are.

After that print the non-beautiful number first with as much zeros
as counted on the rest of the numbers, some exceptions are integrated
in the code.