Strings
Signed-off-by: David Jacobo <jguillen@cimat.mx>

Process once to count the number of times "BA" and "BA" appear, if
one of the other is 0 then the answer is NO, if one of them is one
then you need to check if there is a collition free way to pick them.
ANd if both of their countins are >= 2, then you know right away
there is a solution.