Combinatory, implementation
Set A[0] to all his different 4 combinations, after that, try to
match the rest of the array to this given A[0] (called pivot in code).
There are 4 posible starts, and per each cell you need to evaluate
all the possible combinations, so O(2^4*N).

LINK=https://www.codechef.com/JULY16/problems/CHEFTET